flatten the dataplane test package
