+++ src/integration/java/io/pinecone/integration/dataPlane/UpdateFetchAndQueryPodTest.java
@@ -1,4 +1,4 @@
-package io.pinecone.integration.dataPlane.pod;
+package io.pinecone.integration.dataPlane;
 
 import com.google.protobuf.Struct;
 import com.google.protobuf.Value;
@@ -28,7 +28,7 @@ import static io.pinecone.helpers.BuildUpsertRequest.*;
 import static io.pinecone.helpers.IndexManager.createIndexIfNotExistsDataPlane;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class UpdateFetchAndQueryTest {
+public class UpdateFetchAndQueryPodTest {
     private static PineconeConnection connection;
     private static final int dimension = 3;
